new_admin_panel_sprint_1/02_movies_admin/requirements.txt
django==3.2
flake8==4.0
python-dotenv==0.20
psycopg2-binary==2.9
django-split-settings==1.1
django-debug-toolbar==3.5
